turned CscCount to uint8 according to the latest EF spec

+++ beacon_chain/nimbus_beacon_node.nim
@@ -1154,13 +1154,13 @@ proc addDenebMessageHandlers(
     node.network.subscribe(topic, basicParams)
 
   if node.config.subscribeAllSubnets:
-    node.network.loadCscnetsMetadata(DATA_COLUMN_SIDECAR_SUBNET_COUNT.uint64)
+    node.network.loadCscnetsMetadata(DATA_COLUMN_SIDECAR_SUBNET_COUNT.uint8)
   elif not node.config.subscribeAllSubnets:
     let csc = node.config.custodySubnetCount
     if csc.isSome and csc.get < DATA_COLUMN_SIDECAR_SUBNET_COUNT:
-      node.network.loadCscnetsMetadata(csc.get.uint64)
+      node.network.loadCscnetsMetadata(csc.get.uint8)
     else:
-      node.network.loadCscnetsMetadata(CUSTODY_REQUIREMENT.uint64)
+      node.network.loadCscnetsMetadata(CUSTODY_REQUIREMENT.uint8)
 
 proc addElectraMessageHandlers(
     node: BeaconNode, forkDigest: ForkDigest, slot: Slot) =
