+++ beacon_chain/spec/datatypes/eip7594.nim
@@ -77,7 +77,7 @@ type
     column_index*: ColumnIndex
     row_index*: RowIndex
 
-  CscCount* = uint64
+  CscCount* = uint8
 
   # https://github.com/ethereum/consensus-specs/blob/v1.5.0-alpha.4/specs/_features/eip7594/p2p-interface.md#metadata
   MetaData* = object
